// src/rencana-produksi/interface/rencana-produksi.interface.ts
import { InitialShift } from "@app/initial-shift/initial-shift.entity";
import { Line } from "@app/line/line.entity";
import { User } from "@app/user/user.entity";
import { InitialSku } from "@app/initial-sku/initial-sku.entity";

export interface IRencanaProduksi {
  readonly id?: number;
  readonly po_number?: string;
  readonly standart_ct?: number;
  readonly target_produksi?: number;
  readonly date?: string;
  readonly created_at?: string;
  readonly updated_at?: string;
  readonly deleted_at?: string;

  readonly shift?: InitialShift;
  readonly line?: Line;
  readonly sku?: InitialSku;
  readonly supervisor?: User;
}